src: tidy up types, add necessary casts
Cherry-picked from #13489
Closes #13614

+++ src/tool_writeout.c
@@ -443,7 +443,7 @@ static int writeLong(FILE *stream, const struct writeoutvar *wovar,
       valid = true;
       break;
     case VAR_EXITCODE:
-      longinfo = per_result;
+      longinfo = (long)per_result;
       valid = true;
       break;
     case VAR_URLNUM:
